# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Application Test",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/transactions": {
            "get": {
                "tags": [
                    "transaction"
                ],
                "summary": "Fetch all transactions",
                "description": "Returns all transactions",
                "operationId": "listTransactions",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "An authorization header is required after login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "for search data",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Accounts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/Transaction"
                                                    },
                                                    {
                                                        "properties": {
                                                            "account": {
                                                                "type": "array",
                                                                "items": {
                                                                    "allOf": [
                                                                        {
                                                                            "$ref": "#/components/schemas/Account"
                                                                        }
                                                                    ]
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "transaction"
                ],
                "summary": "Add a new transactions",
                "description": "Adding new transaction data",
                "operationId": "addTransaction",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "An authorization header is required after login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "transaction_date": {
                                        "type": "string"
                                    },
                                    "reference": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "account_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Missing transaction parameter"
                    }
                }
            }
        },
        "/api/admin/transactions/{transactionID}": {
            "get": {
                "tags": [
                    "transaction"
                ],
                "summary": "Get transaction by ID",
                "description": "Returns a single transaction data",
                "operationId": "getTransactionByID",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "An authorization header is required after login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "transactionID",
                        "in": "path",
                        "description": "ID of transaction to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/Transaction"
                                                    },
                                                    {
                                                        "properties": {
                                                            "account": {
                                                                "type": "array",
                                                                "items": {
                                                                    "allOf": [
                                                                        {
                                                                            "$ref": "#/components/schemas/Account"
                                                                        }
                                                                    ]
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "transaction"
                ],
                "summary": "Edit transaction",
                "description": "Updating transaction data",
                "operationId": "editTransaction",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "An authorization header is required after login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "transactionID",
                        "in": "path",
                        "description": "ID of transaction post to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int"
                        }
                    }
                ],
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "transaction_date": {
                                        "type": "string"
                                    },
                                    "reference": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "account_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Missing transaction parameter"
                    },
                    "404": {
                        "description": "transaction not found"
                    },
                    "400": {
                        "description": "Invalid ID transaction"
                    }
                }
            },
            "delete": {
                "tags": [
                    "transaction"
                ],
                "summary": "Delete transaction based on ID",
                "description": "",
                "operationId": "deleteTransactionById",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "An authorization header is required after login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "transactionID",
                        "in": "path",
                        "description": "ID of transaction to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "transaction not found"
                    }
                }
            }
        },
        "/api/admin/transactions/month": {
            "get": {
                "tags": [
                    "transaction"
                ],
                "summary": "Get transaction by month",
                "description": "Returns a single transaction data",
                "operationId": "getTransactionByMonth",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "An authorization header is required after login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/Transaction"
                                                    },
                                                    {
                                                        "properties": {
                                                            "account": {
                                                                "type": "array",
                                                                "items": {
                                                                    "allOf": [
                                                                        {
                                                                            "$ref": "#/components/schemas/Account"
                                                                        }
                                                                    ]
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Account": {
                "title": "Account",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "account_type": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Transaction": {
                "title": "Transaction",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "transaction_date": {
                        "type": "string"
                    },
                    "reference": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "integer"
                    },
                    "account_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        []
    ]
}